refactor(ui): align general and start media ownership
