+++ src/Foundry/Views/GeneralConfigurationPage.xaml
@@ -5,6 +5,7 @@
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
       xmlns:dev="using:DevWinUI"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:wct="using:CommunityToolkit.WinUI.Controls"
       mc:Ignorable="d">
     <ScrollView Margin="{ThemeResource ContentPageMargin}"
                 Padding="{ThemeResource ContentPagePadding}"
@@ -15,20 +16,10 @@
             <TextBlock x:Uid="GeneralConfigurationPage_Title"
                        Style="{ThemeResource FoundryPageTitleTextBlockStyle}" />
 
-            <dev:SettingsCard x:Name="IsoPathCard"
-                              HeaderIcon="{dev:FontIcon GlyphCode=E8A5}">
-                <StackPanel Orientation="Horizontal"
-                            Spacing="{ThemeResource FoundryInlineControlSpacing}">
-                    <TextBox MinWidth="{ThemeResource FoundryWideInputMinWidth}"
-                             Text="{x:Bind ViewModel.IsoOutputPath, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
-                    <Button x:Name="BrowseIsoButton"
-                            MinWidth="{StaticResource SettingActionControlMinWidth}"
-                            Command="{x:Bind ViewModel.BrowseIsoOutputPathCommand}" />
-                </StackPanel>
-            </dev:SettingsCard>
-
-            <dev:SettingsCard x:Name="ArchitectureCard"
-                              HeaderIcon="{dev:FontIcon GlyphCode=E7F4}">
+            <wct:SettingsCard x:Name="ArchitectureCard">
+                <wct:SettingsCard.HeaderIcon>
+                    <FontIcon Glyph="&#xE7F4;" />
+                </wct:SettingsCard.HeaderIcon>
                 <StackPanel Orientation="Horizontal"
                             Spacing="{ThemeResource FoundryInlineControlSpacing}">
                     <ComboBox MinWidth="{StaticResource SettingActionControlMinWidth}"
@@ -38,40 +29,30 @@
                     <ToggleSwitch x:Name="Ca2023Toggle"
                                   IsOn="{x:Bind ViewModel.UseCa2023Signature, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
                 </StackPanel>
-            </dev:SettingsCard>
+            </wct:SettingsCard>
 
-            <dev:SettingsCard x:Name="WinPeLanguageCard">
-                <dev:SettingsCard.HeaderIcon>
+            <wct:SettingsCard x:Name="WinPeLanguageCard">
+                <wct:SettingsCard.HeaderIcon>
                     <FontIcon Glyph="&#xE8AB;" />
-                </dev:SettingsCard.HeaderIcon>
-                <StackPanel Orientation="Horizontal"
-                            Spacing="{ThemeResource FoundryInlineControlSpacing}">
+                </wct:SettingsCard.HeaderIcon>
+                <StackPanel Spacing="{ThemeResource FoundryCompactStackSpacing}">
                     <ComboBox x:Name="WinPeLanguageComboBox"
                               MinWidth="{StaticResource SettingActionControlMinWidth}"
                               IsEnabled="{x:Bind ViewModel.HasWinPeLanguages, Mode=OneWay}"
                               ItemsSource="{x:Bind ViewModel.AvailableWinPeLanguages, Mode=OneWay}"
                               SelectedItem="{x:Bind ViewModel.SelectedWinPeLanguage, Mode=TwoWay}"
                               SelectionChanged="WinPeLanguageComboBox_SelectionChanged" />
+                    <TextBlock Foreground="{ThemeResource TextFillColorSecondaryBrush}"
+                               Text="{x:Bind ViewModel.WinPeLanguageUnavailableDescription, Mode=OneWay}"
+                               TextWrapping="WrapWholeWords"
+                               Visibility="{x:Bind ViewModel.WinPeLanguageUnavailableVisibility, Mode=OneWay}" />
                 </StackPanel>
-            </dev:SettingsCard>
+            </wct:SettingsCard>
 
-            <dev:SettingsCard x:Name="UsbLayoutCard"
-                              HeaderIcon="{dev:FontIcon GlyphCode=E7C3}">
-                <StackPanel Orientation="Horizontal"
-                            Spacing="{ThemeResource FoundryInlineControlSpacing}">
-                    <ComboBox MinWidth="{StaticResource SettingActionControlMinWidth}"
-                              DisplayMemberPath="DisplayName"
-                              ItemsSource="{x:Bind ViewModel.PartitionStyles, Mode=OneWay}"
-                              SelectedItem="{x:Bind ViewModel.SelectedPartitionStyle, Mode=TwoWay}" />
-                    <ComboBox MinWidth="{StaticResource SettingActionControlMinWidth}"
-                              DisplayMemberPath="DisplayName"
-                              ItemsSource="{x:Bind ViewModel.FormatModes, Mode=OneWay}"
-                              SelectedItem="{x:Bind ViewModel.SelectedFormatMode, Mode=TwoWay}" />
-                </StackPanel>
-            </dev:SettingsCard>
-
-            <dev:SettingsExpander x:Name="DriverOptionsCard"
-                                  HeaderIcon="{dev:FontIcon GlyphCode=E90F}">
+            <wct:SettingsExpander x:Name="DriverOptionsCard">
+                <wct:SettingsExpander.HeaderIcon>
+                    <FontIcon Glyph="&#xE90F;" />
+                </wct:SettingsExpander.HeaderIcon>
                 <StackPanel Orientation="Horizontal"
                             Spacing="{ThemeResource FoundryInlineControlSpacing}">
                     <ToggleSwitch x:Name="DellDriversToggle"
@@ -79,8 +60,8 @@
                     <ToggleSwitch x:Name="HpDriversToggle"
                                   IsOn="{x:Bind ViewModel.IncludeHpDrivers, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
                 </StackPanel>
-                <dev:SettingsExpander.Items>
-                    <dev:SettingsCard x:Name="CustomDriverDirectoryCard">
+                <wct:SettingsExpander.Items>
+                    <wct:SettingsCard x:Name="CustomDriverDirectoryCard">
                         <StackPanel Orientation="Horizontal"
                                     Spacing="{ThemeResource FoundryInlineControlSpacing}">
                             <TextBox MinWidth="{ThemeResource FoundryWideInputMinWidth}"
@@ -89,17 +70,20 @@
                                     MinWidth="{StaticResource SettingActionControlMinWidth}"
                                     Command="{x:Bind ViewModel.BrowseCustomDriverDirectoryCommand}" />
                         </StackPanel>
-                    </dev:SettingsCard>
-                </dev:SettingsExpander.Items>
-            </dev:SettingsExpander>
+                    </wct:SettingsCard>
+                </wct:SettingsExpander.Items>
+            </wct:SettingsExpander>
 
-            <dev:SettingsCard x:Name="CreateMediaCard"
-                              HeaderIcon="{dev:FontIcon GlyphCode=E768}">
+            <wct:SettingsCard x:Name="CreateMediaCard">
+                <wct:SettingsCard.HeaderIcon>
+                    <FontIcon Glyph="&#xE768;" />
+                </wct:SettingsCard.HeaderIcon>
                 <Button x:Name="CreateMediaButton"
                         MinWidth="{StaticResource SettingActionControlMinWidth}"
                         IsEnabled="{x:Bind ViewModel.CanCreateMedia, Mode=OneWay}"
+                        Style="{ThemeResource AccentButtonStyle}"
                         Click="CreateMediaButton_Click" />
-            </dev:SettingsCard>
+            </wct:SettingsCard>
         </StackPanel>
     </ScrollView>
 </Page>
