+++ src/Foundry/Views/StartPage.xaml
@@ -5,6 +5,7 @@
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
       xmlns:dev="using:DevWinUI"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:wct="using:CommunityToolkit.WinUI.Controls"
       mc:Ignorable="d">
     <ScrollView Margin="{ThemeResource ContentPageMargin}"
                 Padding="{ThemeResource ContentPagePadding}"
@@ -40,9 +41,25 @@
                 </Grid>
             </Border>
 
-            <dev:SettingsCard x:Name="UsbTargetCard"
-                              Description="{x:Bind ViewModel.UsbCandidateStatus, Mode=OneWay}"
-                              HeaderIcon="{dev:FontIcon GlyphCode=E88E}">
+            <wct:SettingsCard x:Name="IsoPathCard">
+                <wct:SettingsCard.HeaderIcon>
+                    <FontIcon Glyph="&#xE8A5;" />
+                </wct:SettingsCard.HeaderIcon>
+                <StackPanel Orientation="Horizontal"
+                            Spacing="{ThemeResource FoundryInlineControlSpacing}">
+                    <TextBox MinWidth="{ThemeResource FoundryWideInputMinWidth}"
+                             Text="{x:Bind ViewModel.IsoOutputPath, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
+                    <Button x:Name="BrowseIsoButton"
+                            MinWidth="{StaticResource SettingActionControlMinWidth}"
+                            Command="{x:Bind ViewModel.BrowseIsoOutputPathCommand}" />
+                </StackPanel>
+            </wct:SettingsCard>
+
+            <wct:SettingsExpander x:Name="UsbTargetCard"
+                                  Description="{x:Bind ViewModel.UsbCandidateStatus, Mode=OneWay}">
+                <wct:SettingsExpander.HeaderIcon>
+                    <FontIcon Glyph="&#xE88E;" />
+                </wct:SettingsExpander.HeaderIcon>
                 <StackPanel Orientation="Horizontal"
                             Spacing="{ThemeResource FoundryInlineControlSpacing}">
                     <ComboBox MinWidth="{ThemeResource FoundryMediumInputMinWidth}"
@@ -53,10 +70,26 @@
                             MinWidth="{StaticResource SettingActionControlMinWidth}"
                             Command="{x:Bind ViewModel.RefreshUsbCandidatesCommand}" />
                 </StackPanel>
-            </dev:SettingsCard>
+                <wct:SettingsExpander.Items>
+                    <wct:SettingsCard x:Name="UsbPartitionStyleCard">
+                        <ComboBox MinWidth="{StaticResource SettingActionControlMinWidth}"
+                                  DisplayMemberPath="DisplayName"
+                                  ItemsSource="{x:Bind ViewModel.PartitionStyles, Mode=OneWay}"
+                                  SelectedItem="{x:Bind ViewModel.SelectedPartitionStyle, Mode=TwoWay}" />
+                    </wct:SettingsCard>
+                    <wct:SettingsCard x:Name="UsbFormatModeCard">
+                        <ComboBox MinWidth="{StaticResource SettingActionControlMinWidth}"
+                                  DisplayMemberPath="DisplayName"
+                                  ItemsSource="{x:Bind ViewModel.FormatModes, Mode=OneWay}"
+                                  SelectedItem="{x:Bind ViewModel.SelectedFormatMode, Mode=TwoWay}" />
+                    </wct:SettingsCard>
+                </wct:SettingsExpander.Items>
+            </wct:SettingsExpander>
 
-            <dev:SettingsCard x:Name="FinalCommandsCard"
-                              HeaderIcon="{dev:FontIcon GlyphCode=E7C1}">
+            <wct:SettingsCard x:Name="FinalCommandsCard">
+                <wct:SettingsCard.HeaderIcon>
+                    <FontIcon Glyph="&#xE7C1;" />
+                </wct:SettingsCard.HeaderIcon>
                 <StackPanel Orientation="Horizontal"
                             Spacing="{ThemeResource FoundryInlineControlSpacing}">
                     <Button x:Name="CreateIsoButton"
@@ -68,7 +101,7 @@
                             Command="{x:Bind ViewModel.CreateUsbCommand}"
                             IsEnabled="{x:Bind ViewModel.CanCreateUsb, Mode=OneWay}" />
                 </StackPanel>
-            </dev:SettingsCard>
+            </wct:SettingsCard>
         </StackPanel>
     </ScrollView>
 </Page>
